requirements.txt
click
azure-devops
python-dotenv
beautifulsoup4